clippy lints; make pruning mod
